+++ tests/compio_write_limits.rs
@@ -2,7 +2,7 @@
 
 use compio::buf::{BufResult, IoBuf, IoBufMut};
 use compio::io::{AsyncRead, AsyncWrite, util::Splittable};
-use sockudo_ws::{Config, Error};
+use sockudo_ws::Config;
 use std::{cell::Cell, io, rc::Rc};
 
 #[derive(Clone)]
@@ -33,11 +33,7 @@ impl Splittable for RecordingIo {
     }
 }
 fn config() -> Config {
-    Config::builder()
-        .auto_ping(false)
-        .idle_timeout(0)
-        .max_backpressure(8)
-        .build()
+    Config::builder().auto_ping(false).idle_timeout(0).build()
 }
 macro_rules! limit_case {
     ($name:ident, $make:expr) => {
@@ -45,37 +41,45 @@ macro_rules! limit_case {
         async fn $name() {
             let written = Rc::new(Cell::new(0));
             let (mut writer, _guard) = ($make)(RecordingIo(written.clone()));
-            assert!(matches!(
-                writer.send_text("oversized").await,
-                Err(Error::BufferFull)
-            ));
-            assert_eq!(written.get(), 0);
+            writer
+                .send_binary(bytes::Bytes::from(vec![1; 2 * 1024 * 1024]))
+                .await
+                .unwrap();
+            assert_eq!(written.get(), 2 * 1024 * 1024 + 14);
+            writer.send_text("still open").await.unwrap();
+            assert_eq!(written.get(), 2 * 1024 * 1024 + 30);
         }
     };
 }
-limit_case!(unified_rejects_encoded_overflow, |io| (
+limit_case!(unified_allows_default_large_messages, |io| (
     sockudo_ws::CompioWebSocketStream::client(io, config()),
     ()
 ));
-limit_case!(split_rejects_encoded_overflow, |io| {
+limit_case!(split_allows_default_large_messages, |io| {
     let (reader, writer) = sockudo_ws::CompioWebSocketStream::client(io, config()).split();
     (writer, reader)
 });
 #[cfg(feature = "permessage-deflate")]
-limit_case!(compressed_rejects_encoded_overflow, |io| (
+limit_case!(compressed_allows_default_large_messages, |io| (
     sockudo_ws::compio::CompioCompressedWebSocketStream::client(
         io,
         config(),
-        sockudo_ws::DeflateConfig::default()
+        sockudo_ws::DeflateConfig {
+            compression_threshold: usize::MAX,
+            ..Default::default()
+        }
     ),
     ()
 ));
 #[cfg(feature = "permessage-deflate")]
-limit_case!(compressed_split_rejects_encoded_overflow, |io| {
+limit_case!(compressed_split_allows_default_large_messages, |io| {
     let (reader, writer) = sockudo_ws::compio::CompioCompressedWebSocketStream::client(
         io,
         config(),
-        sockudo_ws::DeflateConfig::default(),
+        sockudo_ws::DeflateConfig {
+            compression_threshold: usize::MAX,
+            ..Default::default()
+        },
     )
     .split();
     (writer, reader)